+++ Core/Domain/Activity.cs
@@ -5,7 +5,7 @@ namespace StockTracker.Core.Domain
     public class Activity:BaseObject
     {
 
-        public Activity(int id, DateTime activityDate, float high, float low, float open, float close, int volume) : base()
+        public Activity(int id, DateTime activityDate, decimal high, decimal low, decimal open, decimal close, int volume) : base()
         {
             Id = id;
             ActivityDate = activityDate;
@@ -18,10 +18,10 @@ namespace StockTracker.Core.Domain
 
         //Properties
         //public int Id { get; set; }
-        public float High { get; set; }
-        public float Low { get; set; }
-        public float Open { get; set; }
-        public float Close { get; set; }
+        public decimal High { get; set; }
+        public decimal Low { get; set; }
+        public decimal Open { get; set; }
+        public decimal Close { get; set; }
 
         public int Volume { get; set; }
         //public DateTime ActivityDate { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
