+++ Core/Interfaces/ITradingStructure.cs
@@ -9,8 +9,8 @@ namespace StockTracker.Core.Domain.Interfaces
 
         object GetValue(string PropertyName);
 
-        float GetFloatValue(string PropertyName);
-        void SetFloatValue(string propertyName, float value);
+        decimal GetDecimalValue(string PropertyName);
+        void SetDecimalValue(string propertyName, decimal value);
     }
 
 }
